feat: handle google drive photo links
